aiservices/dietaiservices/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langchain-community==0.0.10
openai==1.6.1
google-generativeai==0.3.2
google-cloud-vision==3.4.5
pillow==10.1.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
aiofiles==23.2.1
pika==1.3.2
aio-pika==9.3.1
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
httpx==0.25.2
asyncio==3.4.3
